Restructure chart generators into a package with an id-keyed registry
src/api/services/charts.py becomes src/api/services/charts/ with one module
per dataset, a base module for the ChartGenerator seam, and an explicit
dataset-id -> generator registry that raises on duplicate registration.
AnalyzeService switches from a first-can_handle-wins scan to a registry
lookup. DATASETS_WITHOUT_CURATED_INSIGHTS is a coverage ratchet enforced by
a unit test: every catalog dataset needs a generator or a deliberate
exclusion entry. Tests move to the mirror-path tests/unit/api/services/charts/.
