+++ src/api/services/charts/__init__.py
@@ -1,0 +1,14 @@
+from src.api.services.charts.base import ChartGenerator, column_to_rows
+from src.api.services.charts.registry import (
+    DATASETS_WITHOUT_CURATED_INSIGHTS,
+    GENERATORS,
+    register,
+)
+
+__all__ = [
+    "ChartGenerator",
+    "column_to_rows",
+    "DATASETS_WITHOUT_CURATED_INSIGHTS",
+    "GENERATORS",
+    "register",
+]
